ncbw-training-portal/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.11
bcrypt==4.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.31
Werkzeug==3.0.1
reportlab==4.0.8
#reportlab==4.2.2

